app/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
slowapi>=0.1.7
azure-storage-blob>=12.18.0
redis>=5.0.0
requests>=2.31.0
